Rename SdrProxy.hpp to SdrTools.hpp, split into source & header
