+++ src/CMakeLists.txt
@@ -203,7 +203,8 @@ set(types_files
     nupic/types/Serializable.hpp
     nupic/types/Sdr.hpp
     nupic/types/Sdr.cpp
-    nupic/types/SdrProxy.hpp
+    nupic/types/SdrTools.hpp
+    nupic/types/SdrTools.cpp
 )
 
 set(utils_files
