+++ src/test/unit/types/SdrTest.cpp
@@ -17,7 +17,7 @@
 
 #include <gtest/gtest.h>
 #include <nupic/types/Sdr.hpp>
-#include <nupic/types/SdrProxy.hpp>
+#include <nupic/types/SdrTools.hpp>
 #include <vector>
 #include <random>
